Add files via upload
LEVEL 3: 
Task 1: Restaurant Reviews
Q1 Analyze the text reviews to identify the most common positive and negative keywords.
Q2 Calculate the average length of reviews and explore if there is a relationship between review length and rating.

Task 2 : Votes Analysis
Q1 Identify the restaurants with the highest and lowest number of votes.
Q2 Analyze if there is a correlation between the number of votes and the rating of a restaurant.

Task 3: Price Range vs. Online Delivery and Table Booking
Q1 Analyze if there is a relationship between the price range and the availability of online delivery and table booking.
Q2 Determine if higher-priced restaurants are more likely to offer these services.